🏗️ 🛠️ refactor: relocate debug and wordwrap managers to the editor directory
