+++ src/keymaps/index.ts
@@ -3,7 +3,7 @@ export { AICommandsManager }        from './ai/commands-manager';
 export { AIToggleManager }          from './ai/toggle-manager';
 export { TerminalManager }          from './terminal/side-panel';
 export { BottomTerminalManager }    from './terminal/bottom-panel';
-export { DebugManager }             from '../debug/panel';
-export { WordWrapManager }          from '../wordwrap/manager';
+export { DebugManager }             from '../editor/debug/panel';
+export { WordWrapManager }          from '../editor/wordwrap/manager';
 export { PanelCommandsManager }     from '../notifications/panels/commands';
 export { recoverSidePanelState }    from './terminal/startup-recovery';
